first edition of base test case
